+++ bnp_na_lib/__init__.py
@@ -1,2 +1,2 @@
-"""Helper modules for bnp_na V13.18."""
-__version__ = "V13.18"
+"""Helper modules for bnp_na V13.19."""
+__version__ = "V13.19"